requirements.txt
Appium-Python-Client==2.11.1
pytest==7.4.2
